RouteResults are now shown on ResultRoutes (seperate Page, design has to be adjusted) Refactored Parts of RouterPlanner
